switch to 32-bit

+++ src/process.c
@@ -28,14 +28,15 @@ void generate_hamming_window_q15(int16_t *window, int frame_size) {
 }
 
 // Aplica a janela Hamming com ponto fixo
-void hamming_window_fixed(int16_t *frame, const int16_t *window_q15, int frame_size) {
+void hamming_window_fixed(int32_t *frame, const int16_t *window_q15, int frame_size) {
     for (int i = 0; i < frame_size; i++) {
         int32_t temp = (int32_t)frame[i] * window_q15[i];
-        frame[i] = (int16_t)(temp >> 15);  // retorna para Q15
+        frame[i] = (temp >> 15);  // retorna para Q15
+        // POSSIVEL ERRO VOLTAR
     }
 }
 
-void hamming_window(int16_t *frame, int frame_size) {
+void hamming_window(int32_t *frame, int frame_size) {
     // Aloca espaço para a janela Hamming em Q15
     int16_t window_q15[frame_size];
 
@@ -47,18 +48,18 @@ void hamming_window(int16_t *frame, int frame_size) {
 }
 
 // Função para criar os frames do sinal
-int16_t** frame_signal_int(const int16_t *samples, int num_samples, int frame_size, int frame_step, int *out_num_frames) {
+int32_t** frame_signal_int(const int16_t *samples, int num_samples, int frame_size, int frame_step, int *out_num_frames) {
     int num_frames = ceil_div((num_samples - frame_size), frame_step) + 1;
 
     // Aloca matriz de frames (num_frames x frame_size)
-    int16_t **frames = (int16_t **)malloc(num_frames * sizeof(int16_t *));
+    int32_t **frames = (int32_t **)malloc(num_frames * sizeof(int32_t *));
     if (!frames) {
         fprintf(stderr, "Erro ao alocar memória para frames.\n");
         return NULL;
     }
 
     for (int i = 0; i < num_frames; i++) {
-        frames[i] = (int16_t *)calloc(frame_size, sizeof(int16_t));  // Zera para padding automático
+        frames[i] = (int32_t *)calloc(frame_size, sizeof(int32_t));  // Zera para padding automático
         if (!frames[i]) {
             fprintf(stderr, "Erro ao alocar memória para o frame %d.\n", i);
             // Libera frames anteriores se der erro
@@ -72,7 +73,7 @@ int16_t** frame_signal_int(const int16_t *samples, int num_samples, int frame_si
         for (int j = 0; j < frame_size; j++) {
             int sample_index = frame_start + j;
             if (sample_index < num_samples) {
-                frames[i][j] = samples[sample_index];
+                frames[i][j] = (int32_t)samples[sample_index];
             } else {
                 frames[i][j] = 0;  // Zero padding
             }
